# Optimized setup.py for version control