This PR optimises pip setup by referencing already cloned files instead of downloading them from https git repo.
It also simplifies version management and installation by removing any repository specific information (branch/tag/commit) from setup.py. Each version of repo (tag/branch/commit) can have its own mapping between setup.py and wheel names.

After merging the same installation method should still be used: 
```
pip install --upgrade pip setuptools
pip install --upgrade git+https://github.com/vmware/vsphere-automation-sdk-python.git`
```

This method was tested across different OS Versions (Linux CentOS/Photon, Mac OS X, Windows). 

Find below the test results:

Notes: 
* With current implementation of PEP 440 in pip/setuptools we still need to `localhost` in path url to have compatibility across platforms.
* It would be good to tag the repository after merging to enable installing consistent version with `pip install --upgrade git+https://github.com/vmware/vsphere-automation-sdk-python.git@<VERSION TAG>`

# MacOS X (venv)

## Python 2
```
~ $ virtualenv -p python vsphere
~ $ cd vsphere
~ $ source ./bin/activate
~ $ pip install --upgrade pip setuptools
~ $ pip install git+https://github.com/khnazaretyan/vsphere-automation-sdk-python.git@fix-setup_req_file
DEPRECATION: Python 2.7 will reach the end of its life on January 1st, 2020. Please upgrade your Python as Python 2.7 won't be maintained after that date. A future version of pip will drop support for Python 2.7.
Collecting git+https://github.com/khnazaretyan/vsphere-automation-sdk-python.git@fix-setup_req_file
  Cloning https://github.com/khnazaretyan/vsphere-automation-sdk-python.git (to revision fix-setup_req_file) to /private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd
  Running command git clone -q https://github.com/khnazaretyan/vsphere-automation-sdk-python.git /private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd
  Running command git checkout -b fix-setup_req_file --track origin/fix-setup_req_file
  Switched to a new branch 'fix-setup_req_file'
  Branch 'fix-setup_req_file' set up to track remote branch 'fix-setup_req_file' from 'origin'.
Collecting lxml>=4.3.0 (from vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/c6/ad/fe41992c0d3f1a8935f35bcfa6ac3f2d2f022959cf2d52651e3a1b432b11/lxml-4.3.4-cp27-cp27m-macosx_10_6_intel.macosx_10_9_intel.macosx_10_9_x86_64.macosx_10_10_intel.macosx_10_10_x86_64.whl
Collecting pyVmomi>=6.7 (from vSphere-Automation-SDK==1.0.0)
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-client-bindings/vapi_client_bindings-3.0.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-common-client/vapi_common_client-2.12.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vmc-client-bindings/vmc_client_bindings-1.6.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/nsx-python-sdk/nsx_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/nsx-policy-python-sdk/nsx_policy_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/nsx-vmc-policy-python-sdk/nsx_vmc_policy_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/nsx-vmc-aws-integration-python-sdk/nsx_vmc_aws_integration_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Collecting suds (from vSphere-Automation-SDK==1.0.0)
Collecting requests>=2.3.0 (from pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/51/bd/23c926cd341ea6b7dd0b2a00aba99ae0f828be89d72b2190f27c11d4b7fb/requests-2.22.0-py2.py3-none-any.whl
Collecting six>=1.7.3 (from pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/73/fb/00a976f728d0d1fecfe898238ce23f502a721c0ac0ecfedb80e0d88c64e9/six-1.12.0-py2.py3-none-any.whl
Collecting pyOpenSSL>=18.0.0 (from vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/01/c8/ceb170d81bd3941cbeb9940fc6cc2ef2ca4288d0ca8929ea4db5905d904d/pyOpenSSL-19.0.0-py2.py3-none-any.whl
Requirement already satisfied: setuptools in ./lib/python2.7/site-packages (from vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0) (41.0.1)
Collecting urllib3!=1.25.0,!=1.25.1,<1.26,>=1.21.1 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/e6/60/247f23a7121ae632d62811ba7f273d0e58972d75e58a94d329d51550a47d/urllib3-1.25.3-py2.py3-none-any.whl
Collecting certifi>=2017.4.17 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/69/1b/b853c7a9d4f6a6d00749e94eb6f3a041e342a885b87340b79c1ef73e3a78/certifi-2019.6.16-py2.py3-none-any.whl
Collecting chardet<3.1.0,>=3.0.2 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/bc/a9/01ffebfb562e4274b6487b4bb1ddec7ca55ec7510b22e4c51f14098443b8/chardet-3.0.4-py2.py3-none-any.whl
Collecting idna<2.9,>=2.5 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/14/2c/cd551d81dbe15200be1cf41cd03869a46fe7226e7450af7a6545bfc474c9/idna-2.8-py2.py3-none-any.whl
Collecting cryptography>=2.3 (from pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/e2/bf/3b641820c561aedde134e88528ba68dffe41ed238899fab7f7ef20118aaf/cryptography-2.7-cp27-cp27m-macosx_10_6_intel.whl
Collecting enum34; python_version < "3" (from cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/c5/db/e56e6b4bbac7c4a06de1c50de6fe1ef3810018ae11732a50f15f62c7d050/enum34-1.1.6-py2-none-any.whl
Collecting cffi!=1.11.3,>=1.8 (from cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/16/f6/46a3dece43541b2cbf3776ec2299e370a2408d9380958401cacb6d101853/cffi-1.12.3-cp27-cp27m-macosx_10_6_intel.whl
Collecting asn1crypto>=0.21.0 (from cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/ea/cd/35485615f45f30a510576f1a56d1e0a7ad7bd8ab5ed7cdc600ef7cd06222/asn1crypto-0.24.0-py2.py3-none-any.whl
Collecting ipaddress; python_version < "3" (from cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/fc/d0/7fc3a811e011d4b388be48a0e381db8d990042df54aa4ef4599a31d39853/ipaddress-1.0.22-py2.py3-none-any.whl
Collecting pycparser (from cffi!=1.11.3,>=1.8->cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-CLYtqd/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
Building wheels for collected packages: vSphere-Automation-SDK
  Building wheel for vSphere-Automation-SDK (setup.py) ... done
  Stored in directory: /private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-ephem-wheel-cache-elKyQl/wheels/1d/b2/fc/850fa96afab399d92e0b370cc4cb104dcc3ac19272219ee778
Successfully built vSphere-Automation-SDK
Installing collected packages: lxml, urllib3, certifi, chardet, idna, requests, six, pyVmomi, enum34, pycparser, cffi, asn1crypto, ipaddress, cryptography, pyOpenSSL, vapi-runtime, vapi-client-bindings, vapi-common-client, vmc-client-bindings, nsx-python-sdk, nsx-policy-python-sdk, nsx-vmc-policy-python-sdk, nsx-vmc-aws-integration-python-sdk, suds, vSphere-Automation-SDK
Successfully installed asn1crypto-0.24.0 certifi-2019.6.16 cffi-1.12.3 chardet-3.0.4 cryptography-2.7 enum34-1.1.6 idna-2.8 ipaddress-1.0.22 lxml-4.3.4 nsx-policy-python-sdk-2.3.0.0.3.13851140 nsx-python-sdk-2.3.0.0.3.13851140 nsx-vmc-aws-integration-python-sdk-2.3.0.0.3.13851140 nsx-vmc-policy-python-sdk-2.3.0.0.3.13851140 pyOpenSSL-19.0.0 pyVmomi-6.7.1.2018.12 pycparser-2.19 requests-2.22.0 six-1.12.0 suds-0.4 urllib3-1.25.3 vSphere-Automation-SDK-1.0.0 vapi-client-bindings-3.0.0 vapi-common-client-2.12.0 vapi-runtime-2.12.0 vmc-client-bindings-1.6.0
```

## Python 3
```
~ $ virtualenv -p python3 vsphere
~ $ cd vsphere
~ $ source ./bin/activate
~ $ pip install --upgrade pip setuptools
~ $ pip install git+https://github.com/khnazaretyan/vsphere-automation-sdk-python.git@fix-setup_req_file
Collecting git+https://github.com/khnazaretyan/vsphere-automation-sdk-python.git@fix-setup_req_file
  Cloning https://github.com/khnazaretyan/vsphere-automation-sdk-python.git (to revision fix-setup_req_file) to /private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam
  Running command git clone -q https://github.com/khnazaretyan/vsphere-automation-sdk-python.git /private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam
  Running command git checkout -b fix-setup_req_file --track origin/fix-setup_req_file
  Switched to a new branch 'fix-setup_req_file'
  Branch 'fix-setup_req_file' set up to track remote branch 'fix-setup_req_file' from 'origin'.
Collecting lxml>=4.3.0 (from vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/1d/16/c220beb03c4b4609924f0fc1bee5e0e5e8082089103fdc0b2a7c8b8bb003/lxml-4.3.4-cp37-cp37m-macosx_10_6_intel.macosx_10_9_intel.macosx_10_9_x86_64.macosx_10_10_intel.macosx_10_10_x86_64.whl
Collecting pyVmomi>=6.7 (from vSphere-Automation-SDK==1.0.0)
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-client-bindings/vapi_client_bindings-3.0.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-common-client/vapi_common_client-2.12.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vmc-client-bindings/vmc_client_bindings-1.6.0-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/nsx-python-sdk/nsx_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/nsx-policy-python-sdk/nsx_policy_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/nsx-vmc-policy-python-sdk/nsx_vmc_policy_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Processing //private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/nsx-vmc-aws-integration-python-sdk/nsx_vmc_aws_integration_python_sdk-2.3.0.0.3.13851140-py2.py3-none-any.whl
Collecting suds-jurko (from vSphere-Automation-SDK==1.0.0)
Collecting six>=1.7.3 (from pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/73/fb/00a976f728d0d1fecfe898238ce23f502a721c0ac0ecfedb80e0d88c64e9/six-1.12.0-py2.py3-none-any.whl
Collecting requests>=2.3.0 (from pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/51/bd/23c926cd341ea6b7dd0b2a00aba99ae0f828be89d72b2190f27c11d4b7fb/requests-2.22.0-py2.py3-none-any.whl
Collecting pyOpenSSL>=18.0.0 (from vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/01/c8/ceb170d81bd3941cbeb9940fc6cc2ef2ca4288d0ca8929ea4db5905d904d/pyOpenSSL-19.0.0-py2.py3-none-any.whl
Requirement already satisfied: setuptools in ./lib/python3.7/site-packages (from vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0) (41.0.1)
Collecting idna<2.9,>=2.5 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/14/2c/cd551d81dbe15200be1cf41cd03869a46fe7226e7450af7a6545bfc474c9/idna-2.8-py2.py3-none-any.whl
Collecting certifi>=2017.4.17 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/69/1b/b853c7a9d4f6a6d00749e94eb6f3a041e342a885b87340b79c1ef73e3a78/certifi-2019.6.16-py2.py3-none-any.whl
Collecting urllib3!=1.25.0,!=1.25.1,<1.26,>=1.21.1 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/e6/60/247f23a7121ae632d62811ba7f273d0e58972d75e58a94d329d51550a47d/urllib3-1.25.3-py2.py3-none-any.whl
Collecting chardet<3.1.0,>=3.0.2 (from requests>=2.3.0->pyVmomi>=6.7->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/bc/a9/01ffebfb562e4274b6487b4bb1ddec7ca55ec7510b22e4c51f14098443b8/chardet-3.0.4-py2.py3-none-any.whl
Collecting cryptography>=2.3 (from pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/63/4e/57b7a6bd98906872fcd2531e74b532de2abe17d675a5cf171931fcb4a9e8/cryptography-2.7-cp34-abi3-macosx_10_6_intel.whl
Collecting cffi!=1.11.3,>=1.8 (from cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/f0/48/5aa4ea664eba26dd5142558d04762f5065c02220b4665b3f7eecb9bb614e/cffi-1.12.3-cp37-cp37m-macosx_10_9_x86_64.whl
Collecting asn1crypto>=0.21.0 (from cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
  Using cached https://files.pythonhosted.org/packages/ea/cd/35485615f45f30a510576f1a56d1e0a7ad7bd8ab5ed7cdc600ef7cd06222/asn1crypto-0.24.0-py2.py3-none-any.whl
Collecting pycparser (from cffi!=1.11.3,>=1.8->cryptography>=2.3->pyOpenSSL>=18.0.0->vapi-runtime@ file://localhost//private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-req-build-9xig1vam/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl->vSphere-Automation-SDK==1.0.0)
Building wheels for collected packages: vSphere-Automation-SDK
  Building wheel for vSphere-Automation-SDK (setup.py) ... done
  Stored in directory: /private/var/folders/q5/8gx0gtcx2jb7fqmb3c310knc0000gn/T/pip-ephem-wheel-cache-jpu5dtoe/wheels/1d/b2/fc/850fa96afab399d92e0b370cc4cb104dcc3ac19272219ee778
Successfully built vSphere-Automation-SDK
Installing collected packages: lxml, six, idna, certifi, urllib3, chardet, requests, pyVmomi, pycparser, cffi, asn1crypto, cryptography, pyOpenSSL, vapi-runtime, vapi-client-bindings, vapi-common-client, vmc-client-bindings, nsx-python-sdk, nsx-policy-python-sdk, nsx-vmc-policy-python-sdk, nsx-vmc-aws-integration-python-sdk, suds-jurko, vSphere-Automation-SDK
Successfully installed asn1crypto-0.24.0 certifi-2019.6.16 cffi-1.12.3 chardet-3.0.4 cryptography-2.7 idna-2.8 lxml-4.3.4 nsx-policy-python-sdk-2.3.0.0.3.13851140 nsx-python-sdk-2.3.0.0.3.13851140 nsx-vmc-aws-integration-python-sdk-2.3.0.0.3.13851140 nsx-vmc-policy-python-sdk-2.3.0.0.3.13851140 pyOpenSSL-19.0.0 pyVmomi-6.7.1.2018.12 pycparser-2.19 requests-2.22.0 six-1.12.0 suds-jurko-0.6 urllib3-1.25.3 vSphere-Automation-SDK-1.0.0 vapi-client-bindings-3.0.0 vapi-common-client-2.12.0 vapi-runtime-2.12.0 vmc-client-bindings-1.6.0
```


# CentOS 7

## Python2

```
docker run -it centos /bin/bash

[root@2416f67c84fa /]# yum install -y epel-release

[root@2416f67c84fa /]# yum install -y python27 python2-pip git
[root@2416f67c84fa /]# pip install --upgrade pip setuptools

[root@2416f67c84fa /]# pip install git+https://github.com/khnazaretyan/vsphere-automation-sdk-python.git@fix-setup_req_file
DEPRECATION: Python 2.7 will reach the end of its life on January 1st, 2020. Please upgrade your Python as Python 2.7 won't be maintained after that date. A future version of pip will drop support for Python 2.7.
Collecting git+https://github.com/khnazaretyan/vsphere-automation-sdk-python.git@fix-setup_req_file
  Cloning https://github.com/khnazaretyan/vsphere-automation-sdk-python.git (to revision fix-setup_req_file) to /tmp/pip-req-build-BgC_Hm
  Running command git clone -q https://github.com/khnazaretyan/vsphere-automation-sdk-python.git /tmp/pip-req-build-BgC_Hm
  Running command git checkout -b fix-setup_req_file --track origin/fix-setup_req_file
  Switched to a new branch 'fix-setup_req_file'
  Branch fix-setup_req_file set up to track remote branch fix-setup_req_file from origin.
Collecting lxml>=4.3.0 (from vSphere-Automation-SDK==1.0.0)
  Downloading https://files.pythonhosted.org/packages/af/31/47cce58942bbf4b8f1c975ec2d1ab52141f7b7cf8cdecb58f25546d2c4fd/lxml-4.3.4-cp27-cp27mu-manylinux1_x86_64.whl (5.6MB)
     |################################| 5.6MB 2.9MB/s
Collecting pyVmomi>=6.7 (from vSphere-Automation-SDK==1.0.0)
  Downloading https://files.pythonhosted.org/packages/71/24/0bb1257b3bc89f7b2facdbad91cc56902d116d649a263c242ef32f73110e/pyvmomi-6.7.1.2018.12.zip (632kB)
     |################################| 634kB 11.1MB/s
Processing ./tmp/pip-req-build-BgC_Hm/lib/vapi-runtime/vapi_runtime-2.12.0-py2.py3-none-any.whl
Processing ./tmp/pip-req-build-BgC_Hm/lib/vapi-client-bindings/vapi_client_bindings-3.0.0-py2.py3-none-any.whl
Processing ./tmp/pip-req-build-BgC_Hm/lib/vapi-common-client/vapi_common_client-2.12.0-py2.py3-none-any.whl
Processing ./tmp/pip-req-build-BgC_Hm/lib/vmc-client-bindings/vmc_client_bindings-1.6.0-py2.py3-none-any.whl
…