@tianhao64 2019-06-23
@khnazaretyan Thank you! This looks really good!
btw, wondering do you happen to know is there any way to have wild card inside the setup.py? To simplify the upgrading/maintenance effort?