@khnazaretyan 2019-06-24
@tianhao64 Sorry for the delay in response and many thanks for the review. Unfortunately I haven't found any way to have wildcards for `file://` in setup.py. In [PEP 440](https://www.python.org/dev/peps/pep-0440/) is only wildcard for versions described. 

Another way to simplify maintenance and versioning, could be to exclude the version specifier completely from whl. files (e.g. vapi_runtime **-2.12.0** -py2.py3-none-any.whl -> **vapi_runtime-py2.py3-none-any.whl**) and have versioning using only git tags in repo. In that case we should change only general SDK version (preferably to the same as git tag) in setup.py with each new version of whl.