@tianhao64 2019-06-24
@khnazaretyan Thanks for the suggestion. 
One concern I have is that, AFAIK, without the version in the filename probably it's not easy for people to find out what's the wheel's version. And that may cause some confusion or misuse.