@khnazaretyan 2019-06-24
@tianhao64 Yes, you're absolutely right. In that case it might be hard to differentiate the versions of the wheels.

It would be good to have the tag of the this version so we can have consistent installations by `tag` instead of `commit`: 
`pip install --upgrade git+https://github.com/vmware/vsphere-automation-sdk-python.git@<TAG>`.