@tianhao64 2019-06-25
@khnazaretyan yes, that makes sense. However, I am facing an issue now. We had previously create a few tags/releases with versions that match vsphere versions originally. Although the bindings follow semantic versioning, the github repo versions don't. The github tags/releases don't mean too much previously as it's not tied to setup.py or pip, which was just a snapshot of the repo. 
Now we have a SDK project with setup.py and I think it makes sense to start fresh with 1.0.0. But I am not sure how to handle the conflict with existing bigger tag versions on the repo. I wonder do you have any recommendations?