@khnazaretyan 2019-06-25
@tianhao64 Yes, you're right, I've also seen the tags. As far as I see, in general, we have 4 options:

1. Start with new versioning/tagging as it's done in setup.py and **delete the old tags**:
    * Pros: clean and fresh start and consistent tagging.
    * Cons: old tags/snapshots will be lost.
    * Actions: delete all old tags from the repo. Add vCenter version in the description of the packet in setup.py.

2. Start with new versioning/tagging as it's done in setup.py **and rename the old tags**:
    * Pros: Snapshots will be preserved but not overlap with versioning
    * Cons: will give clients, that are using tags, the opportunity to use the existing snapshots
    * Actions: rename old tags (e.g. **snap-vCenter-6.8.7**). Add vCenter version in the description of the packet in setup.py.

3. Start with new versioning as it's done in setup.py **with migration phase**:
    * Pros: will give the clients, that are using tags, the opportunity to use the existing snapshots for a period of time.
    * Cons: old tags exist for a specific period of time and should be deleted afterwards.
    * Actions: temporarily leave old tags in the repo with depreciation warning and finally delete all old tags. Add vCenter version in the description of the packet in setup.py.
    
4. Continue following the versioning/tagging according to **vCenter versions**:
   * Pros: good from usability perspective
   * Cons: does not start from scratch (and I agree it makes sense to start SDK versioning from scratch)
   * Actions: add new tag (e.g. v6.8.8) modify version in setup.py to match the tag

For me is the first option fully acceptable, but I don't know how critical are the old tags/snapshots.