@tianhao64 2019-06-26
@khnazaretyan Thank you! These suggestions are great! I also prefer the first option. All bindings are backward compatible and we always recommend users to get the latest released bindings regardless of there vCenter versions, therefore I don't think the old tags are critical. It should be ok to rename them. Again really appreciate your effort and contribution to the project!!